Juggernaut/ai_engine/requirements.txt
flask
flask-cors
ultralytics
google-generativeai
Pillow
python-dotenv
numpy